ST-4518 fix comments
