+++ npl-cli-core/src/main/resources/META-INF/native-image/native-image.properties
@@ -1,6 +1,5 @@
 Args = --no-fallback \
        --install-exit-handlers \
-       --enable-url-protocols=http,https \
        -H:+ReportExceptionStackTraces \
        -H:+UnlockExperimentalVMOptions \
        -O1
